further folder restructuring -- Co-authored-by: Parham Berenjian <ParhamInBinary@users.noreply.github.com>
