+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>QuizIt</title>
-    <link rel="icon" href="./src/assets/images/favicon.png" />
+    <link rel="icon" href="/favicon.png" />
     <link
       href="https://fonts.googleapis.com/css2?family=Dosis:wght@400;500;600;700&family=Poppins:ital,wght@0,400;0,500;0,600;0,800;1,400&display=swap"
       rel="stylesheet"
