+++ src/App.tsx
@@ -3,8 +3,7 @@ import { Flex } from '@chakra-ui/react';
 import { useAppContext } from './AppContext';
 import { MainWrapperStyle, PageWrapperStyle } from './GlobalStyles';
 import { EndScene, HomeScene, LoadingScene, OptionsScene, PlayScene } from './scenes';
-import { Header } from './sharedcomponents/Header';
-import { SceneContainer } from './sharedcomponents/Wrappers/SceneContainer';
+import { Header, SceneContainer } from './shared-components';
 import { Scene } from './utils/types';
 
 export default function App() {
