+++ src/scenes/HomeScene/QuizInputForm.tsx
@@ -12,15 +12,15 @@ import { useFormik } from 'formik';
 import { useRef } from 'react';
 import * as Yup from 'yup';
 
-import { useAppContext } from '../../../AppContext';
-import { homeStrings } from '../../../assets';
-import { QuizInputType, Scene } from '../../../utils/types';
+import { useAppContext } from '../../AppContext';
+import { homeStrings } from '../../assets';
+import { QuizInputType, Scene } from '../../utils/types';
 import {
   FormErrorMessageStyle,
   InputRightElementStyle,
   QuizTopicFormControlStyle,
   SubmitQuizTopicButtonStyle,
-} from '../styles';
+} from './styles';
 
 export const QuizInputForm = () => {
   const { setQuizInput, setScene } = useAppContext();
