--- src/scenes/OptionsScene/components/index.ts
@@ -1,1 +1,0 @@
-export * from './Option';
