+++ src/scenes/OptionsScene/index.ts
@@ -1,6 +1,2 @@
+export * from './Option';
 export * from './OptionsScene';
-export * from './components';
-
-import options from './optionsData.json';
-
-export { options as OptionsData };
