--- src/scenes/OptionsScene/optionsData.json
@@ -1,16 +1,0 @@
-{
-  "options": [
-    {
-      "title": "Type of Quiz",
-      "alternatives": ["Multichoice", "True/False"]
-    },
-    {
-      "title": "Amount of Questions",
-      "alternatives": ["5", "10", "15"]
-    },
-    {
-      "title": "Difficulty",
-      "alternatives": ["Easy", "Medium", "Hard"]
-    }
-  ]
-}
