+++ src/shared-components/Radio/RadioStyles.tsx
@@ -1,6 +1,12 @@
 import { radioAnatomy } from '@chakra-ui/anatomy';
 import { createMultiStyleConfigHelpers } from '@chakra-ui/react';
 
+export const QuizRadioContainerStyle = {
+  w: '100%',
+  maxWidth: { base: '360px', sm: '380px', md: '260px', lg: '360px' },
+  borderRadius: '5px',
+};
+
 const { definePartsStyle, defineMultiStyleConfig } = createMultiStyleConfigHelpers(radioAnatomy.keys);
 
 const baseStyle = definePartsStyle({
