+++ src/shared-components/Radio/index.ts
@@ -1,3 +1,2 @@
 export * from './Radio';
 export * from './RadioStyles';
-export * from './getRadioBackgroundColor';
